rqe_iterators: add build_term_iterator
Move the TermIndexReader encoding dispatch from iterators_ffi into rqe_iterators
and add build_term_iterator, a reusable constructor for a term-query iterator over
an in-memory inverted index.

Will be used to create term iterator in eval.rs without using the FFI
API.
